+++ plugins/auth-backend/src/providers/index.ts
@@ -15,6 +15,7 @@
  */
 
 export * from './google';
+export * from './microsoft';
 export { factories as defaultAuthProviderFactories } from './factories';
 
 // Export the minimal interface required for implementing a