+++ plugins/kafka-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-kafka-backend
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@0.9.0
+  - @backstage/backend-common@0.8.5
+
 ## 0.2.7
 
 ### Patch Changes