+++ plugins/kubernetes-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-kubernetes-backend
 
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@0.9.0
+  - @backstage/backend-common@0.8.5
+  - @backstage/plugin-kubernetes-common@0.1.2
+
 ## 0.3.8
 
 ### Patch Changes